Move YSort into Camera Module
- Also add camera plugin to remove camera system scheduling from
main.rs.
